Update imageFeed.js
This is a aewrite of the Image Feed system based on my previous abandoned pull request.

The feed:
- lets the user filter images based on source node. 
- groups images from the same run into batches with a visible marker.

Additionally, the slider for height as well as two new controls have been moved to the ComfyUI menu.

Locking the feed to the left or right has not yet been implemented. Known 'image nodes' are defined in an array, with unrecognised nodes caught via a catch fall filter setting.