Fix lightbox loading position
Issue
When scrolling through images in the Image Feed, there is a brief jitter as images are loaded.

Cause
The loading spinner is an inline sibling element of the image element.

Solution
Use absolute positioning on the spinner root element, relative to the containing pysssss-lightbox-link element.  Use flex to centre the lightbox contents.